+++ front-end/src/Pages/Profile.js
@@ -1,12 +1,12 @@
 import React from "react";
-import Projects from "./Projects";
+import ProjectList from "../Components/ProjectList";
 
 function Profile() {
   //the projects will be passed props of what users profile we are on
   return (
     <div>
       <h1>Profile Page</h1>
-      <Projects />
+      <ProjectList />
     </div>
   );
 }
